MCP Servers/firestore_mcp/requirements.txt
google-cloud-firestore>=2.13.0
mcp>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.25.0
google-cloud-logging>=3.8.0
structlog>=23.1.0